Lathe.models — Module

|====== Lathe.models =====

|____________/ Accessories ___________

|_____models.Pipeline([steps])

|____________/ Continuous models ___________

|_____models.meanBaseline(y)

|_____models.LinearRegression(x,y)

|_____models.LeastSquare(x,y,Type)

|_____models.PowerLog(prob1,prob2)

|____________/ Categorical Models ___________

|_____models.LogisticRegression(x,y)

|_____models.majBaseline(y)

Lathe.models.LeastSquare — Method

Least Squares is ideal for predicting continous features. Many models use Least Squares as a base to build off of.

x = [7,6,5,6,5]

y = [3.4.5.6.3]

xtrain = [7,5,4,5,3,5,7,8]

Type = :LIN

model = Lathe.models.LeastSquare(x,y,Type)

y_pred = Lathe.models.predict(model,xtrain)

HYPER PARAMETERS

Type <- Type determines which Linear Least Square algorithm to use, :LIN, :OLS, :WLS, and :GLS are the three options.

:LIN = Linear Least Square Regression

:OLS = Ordinary Least Squares

:WLS = Weighted Least Squares

:GLS = General Least Squares

==Functions==

predict(xt) <- Returns a prediction from the model based on the xtrain value passed (xt)

Lathe.models.LinearRegression — Method

Linear Regression is a well-known linear function used for predicting continuous features with a mostly linear or semi-linear slope.

==PARAMETERS==

[y] <- Fill with your trainY values. Should be an array of shape (0,1) or (1,0)

[x] <- Fill in with your trainX values. Should be an array of shape (0,1) or (1,0)

==Functions==

predict(xt) <- Returns a prediction from the model based on the xtrain value passed (xt)

Lathe.models.LogisticRegression — Function

Majority class baseline is used to find the most often interpreted classification in an array.

==PARAMETERS==

[X] <- Fill with your trainX values. Should be an array of shape (0,1) or (1,0)

[y] <- Fill with your trainy values. Should be an array of shape (0,1) or (1,0)

λ = .0001 <- Lambda Value

fit_intercept = true <- Boolean determines whether to fit an intercept.

max_iter = 1000 <- Determines the maximum number of iterations for the model to perform.

==Functions==

predict(xt) <- Returns a prediction from the model based on the xtrain value passed (xt)

Lathe.models.MeanBaseline — Method

A mean baseline is great for getting a basic accuracy score in order to make a valid direction for your model.

==PARAMETERS==

[y] <- Fill with your trainY values. Should be an array of shape (0,1) or (1,0)

pipl = Pipeline([StandardScalar(),LinearRegression(trainX,trainy)])

==Functions==

predict(xt) <- Returns a prediction from the model based on the xtrain value passed (xt)

Lathe.models.Network — Method

 'Network' specifically constructs a convolutional nueral network,

 though this network is still being worked on in both methodology

 and in a physical manner,

thanks for understanding.

Lathe.models.Pipeline — Method

Pipelines can contain a predictable Lathe model with preprocessing that occurs automatically. This is done by putting X array processing methods into the iterable steps, and then putting your Lathe model in.

==PARAMETERS==

[steps] <- An iterable list of methods to call for X modification. These mutations should have ALREADY BEEN MADE TO THE TRAIN X.

pipl = Pipeline([StandardScalar(),LinearRegression(trainX,trainy)])

==Functions==

predict(xt) <- Returns a prediction from the model based on the xtrain value passed (xt)

Lathe.models.PowerLog — Method

A powerlog can be used to perform a one-tailed test, as well as get the proper sample size for a testing population.

==PARAMETERS==

p1 <- A Float64 percentage representing the probability of scenario one.

p2 <- A Float64 percentage representing the probability of scenario two. These two probability values should follow these guidelines: p1 = p1 + x = p2

alpha = 0.05 <- Sets an alpha value

Returns power, sample_size

Lathe.models.RidgeRegression — Method

Ridge Regression is another regressor ideal for predicting linear, continuous features.

array = [5,10,15]

scaled_feature = Lathe.preprocess.OneHotEncode(array)

Lathe.models.majClassBaseline — Method

Majority class baseline is used to find the most often interpreted classification in an array.

==PARAMETERS==

[y] <- Fill with your trainY values. Should be an array of shape (0,1) or (1,0)

==Functions==

predict(xt) <- Returns a prediction from the model based on the xtrain value passed (xt)

counts() <- Returns a dictionary with the counts of all inserted keys.

highest() <- Will return a Dictionary key with the count as well as the value for the most interpreted classification.