Lathe.preprocess — Module

|====== Lathe.preprocess =====

|____________/ Generalized Processing ___________

|_____preprocess.TrainTestSplit(array)

|_____preprocess.SortSplit(array)

|_____preprocess.UniformSplit(array)

|____________/ Feature Scaling ___________

|_____preprocess.Rescalar(array)

|_____preprocess.ArbitraryRescale(array)

|_____preprocess.MeanNormalization(array)

|_____preprocess.StandardScalar(array)

|____________/ Categorical Encoding ___________

|_____preprocess.OneHotEncode(array)

Lathe.preprocess.ArbitraryRescale — Method

Arbitrary Rescaling scales a feature based on the minimum and maximum of the array.

array = [5,10,15]

scaled_feature = Lathe.preprocess.Rescalar(array)

==Functions==

predict(xt) <- Returns a prediction from the model based on the xtrain value passed (xt)

Lathe.preprocess.MeanNormalization — Method

Mean Normalization normalizes the data based on the mean.

array = [5,10,15]

scaled_feature = Lathe.preprocess.MeanNormalization(array)

==Functions==

predict(xt) <- Returns a prediction from the model based on the xtrain value passed (xt)

Lathe.preprocess.OneHotEncode — Method

One hot encoder replaces a single feature with sub arrays containing boolean values (1 or 0) for each individual category.

df = DataFrame(:A => ['w','b','w'], :B => [5, 10, 15])

scaled_feature = Lathe.preprocess.OneHotEncode(df,:A)

Lathe.preprocess.Rescalar — Method

Rescalar scales a feature based on the minimum and maximum of the array.

array = [5,10,15]

scaled_feature = Lathe.preprocess.Rescalar(array)

==Functions==

predict(xt) <- Returns a prediction from the model based on the xtrain value passed (xt)

Lathe.preprocess.SortSplit — Function

SortSplit sorts the data from least to greatest, and then splits it, ideal for quartile calculations.

array = [5,10,15]

top25, lower75 = Lathe.preprocess.SortSplit(array,at = 0.75,rev = false)

PARAMETERS:

at:: Percentage value used to determine a point to split the data.

rev:: Reverse, false by default, determines whether to sort least to greatest, or greatest to least.

Lathe.preprocess.Standardize — Method

Standard Scalar z-score normalizes a feature.

array = [5,10,15]

scaled_feature = Lathe.preprocess.StandardScalar(array)

==Functions==

predict(xt) <- Returns a prediction from the model based on the xtrain value passed (xt)

Lathe.preprocess.TrainTestSplit — Method

TrainTestSplit takes either a DataFrame or an Array and splits it according to the at parameter.

[data] <- Iterable dictionary, dataframe, or Array.

a <- Percentage value used to determine a point to split the data.

Lathe.preprocess.UniformSplit — Function

Uniform Split does the exact same thing as ArraySplit(), but observations are returned split, but unsorted and unshuffled.

array = [5,10,15]

test, train = Lathe.preprocess.UniformSplit(array,at = 0.75)

PARAMETERS:

at:: Percentage value used to determine a point to split the data.

Lathe.preprocess.UnitLScale — Method

FUNCTION NOT YET WRITTEN

Unit L Scaling uses eigen values to normalize the data.

array = [5,10,15]

scaled_feature = Lathe.preprocess.UnitLScale(array)