Lathe.stats — Module

|====== Lathe.stats ======

| ~~~~~~~~~~ Base ~~~~~~~~~~~

|_____stats.mean(array)

|_____stats.mode(array)

|_____stats.variance(array)

|_____stats.confiints(data,confidence = .95)

|_____stats.standarderror(array)

|_____stats.std(data)

|_____stats.correlationcoeff(x,y)

|_____stats.z(array)

|_____stats.firstquar(array)

|_____stats.secondquar(array)

|_____stats.thirdquar(array)

| ~~~~~~~~~~ Inferential ~~~~~~~~~~~

|_____stats.independent_t(sample,general)

|_____stats.spearman(var1,var2)

|_____stats.pearson(x,y)

|_____stats.chisqu(array)

|_____stats.sign(array)

|_____stats.f_test(sample,general)

| ~~~~~~~~~~ Bayesian ~~~~~~~~~~~

|_____stats.bay_ther(p,a,b)

|_____stats.cond_prob(p,a,b)

| ~~~~~~~~~~ Distributions ~~~~~~~~~~~

|====== validate ======

|____________/ Metrics ___________

|_____validate.mae(actual,pred)

|_____validate.r2(actual,pred)

Lathe.stats.bernoulli_dist — Method

Function Not yet written

Lathe.stats.binomial_dist — Method

Binomial Distribution is a distribution well known for its use in statistical tests and decision making models.

array = [5,10,15]

r = Lathe.stats.anova(array)

Lathe.stats.chi_dist — Method

FUNCTION NOT YET WRITTEN

Chi Distribution in another well-known distribution well known for being used in statistical tests.

array = [5,10,15]

r = Lathe.stats.anova(array)

Lathe.stats.chisq — Method

Returns a probability using a chi squared distribution.

var1 = [5,10,15]

var2 = [5,4,3,5,6]

p = Lathe.stats.chisq(var1,var2)

Lathe.stats.cond_prob — Method

Performs Bayesian Conditional Probability and returns probability.

prob = .50

prior = .20

evidence = .30

p = Lathe.stats.cond_prob(prob,prior,evidence)

Lathe.stats.confints — Function

Returns the confidence intervals of given data.

array = [5,10,15]

confidence = .98

low, high = Lathe.stats.confints(array,confidence)

PARAMETERS

confidence: Confidence is a float percentage representing the level of confidence required in your test. The confidence metric is used exclusively for calculating the interval.

Lathe.stats.correlationcoeff — Method

Calculates the Correlation Coeffiecient of between two features

x = [5,10,15]

y = [5,10,15]

r = Lathe.stats.correlationcoeff(x,y)

Lathe.stats.f_test — Method

An F test returns a probability of correlation, and is used similarly to a T test.

array = [5,10,15]

r = Lathe.stats.anova(array)

Lathe.stats.firstquar — Method

Returns the point in an array located at 25 percent of the sorted data.

array = [5,10,15]

q1 = Lathe.stats.firstquar(array)

Lathe.stats.getranks — Function

Ranks indices in an array based on quantitative weight (count of the numbers) and returns a new array of the ranks of each column. This function is made primarily for the Wilcox Rank-Sum test.

array = [5,10,15]

q2 = Lathe.stats.secondquar(array)

Lathe.stats.independent_t — Method

Performs an independent T test.

sample = [5,10,15] general = [15,25,35]

t = Lathe.stats.independent_t(sample,general)

Lathe.stats.mae — Method

Mean absolute error (MAE) subtracts two arrays and averages the difference.

array = [5,10,15]

r = Lathe.stats.anova(array)

Lathe.stats.mean — Method

Calculates the mean of a given array.

array = [5,10,15]

mean = Lathe.stats.mean(array)

println(mean)

10

Lathe.stats.median — Method

Calculates the median (numerical center) of a given array.

array = [5,10,15]

median = Lathe.stats.median(array)

println(median)

10

Lathe.stats.mode — Method

Gives the digit most common in a given array

array = [5,10,15,15,10,5,10]

mode = Lathe.stats.mode(array)

println(mode)

10

Lathe.stats.mse — Method

Mean Square error (MSE) subtracts two arrays, squares the difference, and averages the result

array = [5,10,15]

r = Lathe.stats.anova(array)

Lathe.stats.normal_dist — Method

Standard Scalar z-score normalizes a feature.

array = [5,10,15]

scaled_feature = Lathe.preprocess.StandardScalar(array)

Lathe.stats.paired_t — Method

THIS FUNCTION IS NOT YET WRITTEN

Paired T (Dependent T) is a T-test that doesn't require a sample.

array = [5,10,15]

q2 = Lathe.stats.paired_t(var1,var2)

Lathe.stats.pearson — Method

Returns a probability using a Pearson correlation.

x = [5,10,15]

y = [5,4,3,5,6]

r = Lathe.stats.spearman(x,y)

Lathe.stats.r2 — Method

R squared is the correlation coefficient of regression, and is found by squaring the correlation coefficient.

array = [5,10,15]

r = Lathe.stats.anova(array)

Lathe.stats.secondquar — Method

Returns the point in an array located at 50 percent of the sorted data. The second quartile is also known as the median, or the middle of the sorted data.

array = [5,10,15]

q2 = Lathe.stats.secondquar(array)

Lathe.stats.sign — Method

The Sign test determines correlation through negative and positive placement with binomial distribution.

var1 = [5,10,15]

var2 = [19,25,30]

p = Lathe.stats.sign(var1,var2)

Lathe.stats.spearman — Method

Returns a probability using a Spearman correlation.

var1 = [5,10,15]

var2 = [5,4,3,5,6]

r = Lathe.stats.spearman(var1,var2)

Lathe.stats.std — Method

Calculates the Standard Deviation of a given array.

array = [5,10,15]

std = Lathe.stats.std(array)

Lathe.stats.ste — Method

Calculates the Standard Error of an array.

array = [5,10,15]

ste = Lathe.stats.standarderror(array)

Lathe.stats.thirdquar — Method

Returns the point in an array located at 75 percent of the sorted data.

array = [5,10,15]

q2 = Lathe.stats.secondquar(array)

Lathe.stats.variance — Method

Gives the variance of an array..

array = [5,10,15]

variance = Lathe.stats.variance(array)

Lathe.stats.wilcoxrs — Method

FUNCTION NOT YET WRITTEN

Wilcox Sum Rank Tests are used to determine a probability with ranks

var1 = [5,10,15]

var2 = [19,25,30]

p = Lathe.stats.wilcoxsr(var1,var2)

Lathe.stats.wilcoxsr — Method

FUNCTION NOT YET WRITTEN

Wilcox Sum Rank Tests are used to determine a probability with ranks

var1 = [5,10,15]

var2 = [19,25,30]

p = Lathe.stats.wilcoxsr(var1,var2)

Lathe.stats.z — Method

Calculates the Z score of a given array.

array = [5,10,15]

z = Lathe.stats.z(array)